Add files via upload
This adds an incomplete version of the program.  I have finished the enchanting room and the basement and gone back to add comments and documentation to help make my code easier to understand.  
My next tasks are the alchemy room, as well as adding the contents of a piece of paper the user can find and allow the user to read it whenever the like.