Add files via upload
replaces one version of silly_dungeon.py with another.  This new file removes the global variable for inventory, and instead uses a class called "user" and the instance of this class called "player" in it's place.  I also added player specific actions such as read or check inventory in that class.