Add files via upload
Replacing old files.  Made good progress by switching to returning string values in order to change rooms, then using a while loop with an if statement in order to use the function of a specific, previously created instance (ex: using the alchemy variable instead of creating a new Alchemy() room every time).  I could also feed the specific instance of the User class (player in this case) into the room in order to keep the same instances working.  
HOWEVER, the item instances were not properly passed to the room.enter() functions, even though I passed the item instances into the room instances.  Saving this stage of the project to try a different approach. 

I have just now learned that when I simply state Room(), it automatically creates an instance, even if I don't assign it a variable.  So I'll be creating a dictionary that uses strings to look up members of the item class.  Then I will get those values in order to use the Item() functions and access those attributes. Wish me luck!