Add files via upload
Changed back to using lists of strings to manipulate inventories, then use those strings as the key to the all_items dictionary, where the value is the instance of the class items which can then be used. 
Next task: take the "enter" function and condense it to only one main loop, located in the main Room class under the enter funtion.